// src/app/types/opportunities.ts
export interface Opportunity {
  id: number;
  created_at: string | null;
  updated_at: string | null;
  id_contact: string | null;
  full_name: string | null;
  email: string | null;
  phone: string | null;
  opportunity_name: string | null;
  opportunity_st: string | null;
  date_created: string | null;
  tipo_campaña: string | null;
  stage_booked: string | null;
  pipeline_name: string | null;
  pipeline_stage: string | null;
  location: string | null;
  id_opportunity: string | null;
  location_id: string | null;
  pipeline_id: string | null;
  current_st: string | null;
}